+++ lib/wasi/src/syscalls/mod.rs
@@ -9,7 +9,8 @@ use self::types::*;
 use crate::{
     ptr::{Array, WasmPtr},
     state::{
-        host_file_type_to_wasi_file_type, Fd, InodeVal, Kind, WasiFile, WasiState, MAX_SYMLINKS,
+        get_stat_for_kind, host_file_type_to_wasi_file_type, Fd, InodeVal, Kind, WasiFile,
+        WasiState, MAX_SYMLINKS,
     },
     ExitCode,
 };
@@ -187,7 +188,10 @@ pub fn clock_time_get(
     precision: __wasi_timestamp_t,
     time: WasmPtr<__wasi_timestamp_t>,
 ) -> __wasi_errno_t {
-    debug!("wasi::clock_time_get");
+    debug!(
+        "wasi::clock_time_get clock_id: {}, precision: {}",
+        clock_id, precision
+    );
     let memory = ctx.memory(0);
 
     let out_addr = wasi_try!(time.deref(memory));
@@ -791,6 +795,7 @@ pub fn fd_readdir(
         for entry in entries.iter().skip(cookie as usize) {
             cur_cookie += 1;
             let entry_path = entry.path();
+            let entry_path = wasi_try!(entry_path.file_name().ok_or(__WASI_EIO));
             let entry_path_str = entry_path.to_string_lossy();
             let namlen = entry_path_str.len();
             debug!("Returning dirent for {}", entry_path_str);
@@ -1082,7 +1087,7 @@ pub fn path_create_directory(
         entries: Default::default(),
     };
     let new_inode = state.fs.inodes.insert(InodeVal {
-        stat: __wasi_filestat_t::default(),
+        stat: wasi_try!(get_stat_for_kind(&kind).ok_or(__WASI_EIO)),
         is_preopened: false,
         name: path_vec[0].clone(),
         kind,
@@ -1226,10 +1231,7 @@ pub fn path_filestat_get(
                 }
                 let final_path_metadata =
                     wasi_try!(cumulative_path.metadata().map_err(|_| __WASI_EIO));
-                __wasi_filestat_t {
-                    st_filetype: host_file_type_to_wasi_file_type(final_path_metadata.file_type()),
-                    ..Default::default()
-                }
+                wasi_try!(get_stat_for_kind(&state.fs.inodes[inode].kind).ok_or(__WASI_EIO))
             }
         }
         _ => {
@@ -1485,10 +1487,11 @@ pub fn path_open(
                 .fs
                 .create_fd(fs_rights_base, fs_rights_inheriting, fs_flags, child))
         } else {
-            let file_metadata = wasi_try!(file_path.metadata().map_err(|_| __WASI_ENOENT));
+            debug!("Attempting to load file from host system");
+            let file_metadata = file_path.metadata();
             // if entry does not exist in parent directory, try to lazily
             // load it; possibly creating or truncating it if flags set
-            let kind = if file_metadata.is_dir() {
+            let kind = if file_metadata.is_ok() && file_metadata.unwrap().is_dir() {
                 // special dir logic
                 Kind::Dir {
                     parent: Some(cur_dir_inode),
@@ -1533,7 +1536,7 @@ pub fn path_open(
 
             // record lazily loaded or newly created fd
             let new_inode = state.fs.inodes.insert(InodeVal {
-                stat: __wasi_filestat_t::default(),
+                stat: wasi_try!(get_stat_for_kind(&kind).ok_or(__WASI_EIO)),
                 is_preopened: false,
                 name: file_name.clone(),
                 kind,