+++ lib/wasi/tests/wasitests/file_metadata.rs
@@ -1,0 +1,9 @@
+#[test]
+fn test_file_metadata() {
+    assert_wasi_output!(
+        "../../wasitests/file_metadata.wasm",
+        "file_metadata",
+        vec![],
+        "../../wasitests/file_metadata.out"
+    );
+}