+++ src/bin/wasmer.rs
@@ -97,7 +97,11 @@ struct Run {
     #[structopt(long = "dir", multiple = true, group = "wasi")]
     pre_opened_directories: Vec<String>,
 
-    // Custom code loader
+    /// Map a host directory to a different location for the wasm module
+    #[structopt(long = "mapdir", multiple = true)]
+    mapped_dirs: Vec<String>,
+
+    /// Custom code loader
     #[structopt(
         long = "loader",
         raw(possible_values = "LoaderName::variants()", case_insensitive = "true")
@@ -227,6 +231,31 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
     #[cfg(not(target_os = "windows"))]
     let disable_cache = options.disable_cache;
 
+    let mapped_dirs = {
+        let mut md = vec![];
+        for entry in options.mapped_dirs.iter() {
+            if let &[alias, real_dir] = &entry.split(':').collect::<Vec<&str>>()[..] {
+                let pb = PathBuf::from(&real_dir);
+                if let Ok(pb_metadata) = pb.metadata() {
+                    if !pb_metadata.is_dir() {
+                        return Err(format!(
+                            "\"{}\" exists, but it is not a directory",
+                            &real_dir
+                        ));
+                    }
+                } else {
+                    return Err(format!("Directory \"{}\" does not exist", &real_dir));
+                }
+                md.push((alias.to_string(), pb));
+                continue;
+            }
+            return Err(format!(
+                "Directory mappings must consist of two paths separate by a colon. Found {}",
+                &entry
+            ));
+        }
+        md
+    };
     let wasm_path = &options.path;
 
     let mut wasm_binary: Vec<u8> = read_file_contents(wasm_path).map_err(|err| {
@@ -416,6 +445,9 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
             .instantiate(&import_object)
             .map_err(|e| format!("Can't instantiate module: {:?}", e))?;
 
+        if !mapped_dirs.is_empty() {
+            eprintln!("WARN: mapdir is not implemented for emscripten targets");
+        }
         wasmer_emscripten::run_emscripten_instance(
             &module,
             &mut instance,
@@ -445,6 +477,7 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
                     .map(|(k, v)| format!("{}={}", k, v).into_bytes())
                     .collect(),
                 options.pre_opened_directories.clone(),
+                mapped_dirs,
             );
 
             let instance = module